# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inter Finder API",
    "version": "1.0.0",
    "description": "API Documentation for the Inter Finder web application. This provides a clear contract for frontend and backend teams."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/interviews": {
      "post": {
        "summary": "Schedule a new interview",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "applicationId",
                  "date",
                  "time"
                ],
                "properties": {
                  "applicationId": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "time": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interview scheduled successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (company only)"
          },
          "404": {
            "description": "Application not found"
          }
        }
      }
    },
    "/interviews/company": {
      "get": {
        "summary": "Get all interviews for company's applications",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved interviews"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (company only)"
          }
        }
      }
    },
...
    "/interviews/{id}": {
      "put": {
        "summary": "Update interview details",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Interview ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "time": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "scheduled",
                      "completed",
                      "cancelled",
                      "rescheduled"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Interview updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized to update this interview"
          },
          "404": {
            "description": "Interview not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an interview",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Interview ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Interview deleted successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (company only)"
          },
          "404": {
            "description": "Interview not found"
          }
        }
      }
    },
...
    "/stats/interviews": {
      "get": {
        "summary": "Get interview statistics",
        "tags": [
          "Stats"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved interview statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          },
                          "count": {
                            "type": "number"
                          },
                          "avgScheduledDays": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized to view these stats"
          }
        }
      }
    },
...
  },
...
}